RavenDB-6437: * Support O_LARGEFILE (and PerPlatformValues adjust to the arm32) * Use lseek as a fallback to fallocate * Fix lseek usage / err handling * Handle unknown mount type

+++ src/Voron/Platform/Posix/PosixJournalWriter.cs
@@ -63,7 +63,7 @@ namespace Voron.Platform.Posix
                 length = journalSize;
                 try
                 {
-                    PosixHelper.AllocateFileSpace(options, _fd, (ulong) journalSize,filename);
+                    PosixHelper.AllocateFileSpace(options, _fd, journalSize,filename);
                 }
                 catch (Exception)
                 {
